CogVM source as per VMMaker.oscog-eem.2285
Slang:
Fix sources by backing out of the inlining optimization used in the logging
methods that are now unused.  The optmization was careless, not assuming calls
could have side-effects, so in primitiveSubtract the calls to stackIntegerValue
were pushed past the check for primitive failure, and hence
SmallInteger - SmallFloat (in EllipseMidpointTracer>>stepInY) didn't fail.

Spur Cogit:
fix primStringReplace in the JIT.

Plugins:
BitBltSimulation>>copyLoop:
Fix type inference of variable unskew in BitBlt >> copyLoop
Strength reduce the loop split for combinationRule = 3 so that less tests are
done in the inner loop and so trhat when combinstionRule = 3 mergeFn is never
used.  Use an explicit type declaration for unskew instead of trickier relying
on type inference.  Eliminate bogus use of #== for numeric comparison.

Fix definitions for CamerPlugin on win32 by nuking bogus intermediate include
file (cameraOps.h), and explicitly on iOS.

Scripts:
Have the mac makeem scripts for the processor plugins call conf.COG if it
appears not to have been run.