VMMaker.oscog-eem.2286 Simulation on 64-bit Squeak.
Update DoubleWordArray with the low level accessors.  Extend MemoryTests to test these accessors in DoubleWordArray.  Beef up the tests to test all the classes that support the accessors more thoroughly.  Add 64-bit access in Bitmap (which is big endian) as the expected failure.

Have ByteArray use integerAt:[put:]size:signed: for 32 and 64 bit access as these primitives are fast.

Change the oop range in CurrentImageCoInterpreterFacadeFor64BitSpurObjectRepresentation so that it matches code produced by the real JIT on Mac OS X.  Fix a bug in the oop<->object mapping (actually its artguably a bug in IdentityDictionary).  Add IdentityDictionary>>keyAtNonIdentityValue:ifAbsent: to fix the bug.

Slang generation:
Extract logging to a method and have it include the target directory.