Cog VM source as per VMMaker.oscog-eem.2287
Spur:
Rewrite determining the accessorDepth for the meta primitives
(primitiveDoPrimitiveWithArgs and primitiveDoNamedPrimitiveWithArgs, 118 & 218).
Have primitiveDoNamedPrimitiveWithArgs fetch the accessorDepth.  Discard the
confused setting of newMethod to the prim index.  Introduce metaAccessorDepth
and eliminate tempOop3.  This fixes a crash when stepping through something
likely to fail such as [DoubleWordArray new: 64*1024*1024] repeat.

Fix a slip in printDecodeMethodHeaderOop: