Get the StackInterpreterSimulator to a state where it can correctly simulate the DoubleByteArray, WordArray, DoubleWordArray and MemoryTests.  Principally allow the SqueakFFIPrims plugin (ThreadedFFIPlugin) to load and make primitiveFFIIntegerAt[Put] function correctly in simulation.  Change the two primitives to use unalignedShortAt:[put:], unalignedLong32At:[put:] & unalignedLong64At:[put:] and implement these in SpurMemoryManager (ObjectMemory can wait) and have the preambleCCode map these to the original shortAt[put],  long32At[put] & long32At[put] C functions/macros.
Fix a bug in the simulation of LargeIntegersPlugin>>cDigitSub:len:with:len:into: (Nicolas, there may be similar signedness issues in other functions.  If you have time and energy please consider taking a look at the code in the simulator).

Fix a slip in the range comparison for 64-bit integer arguments in genPrimitiveAtPut.