Fix -textenc option parsing in unix vm
The terminal null character of text encoding was not copied to the uppercase variant.
This is a bug, alloca does not set the allocated bytes to zero.

While at it, use VMOPTION macro where we handle vm option strings.
We may get rid of VMOPTION macro in the future, but this is another decision.