CogVM source as per VMMaker.oscog-eem.2314
Cogit:
Fix bad bug in ceSend:above:to:numArgs: which was written to accept an association,
not a class.  The rewrite agrees with the interpreter's definition.

Fix bad bug in ceSend:super:to:numArgs: and ceSend:above:to:numArgs:.  Can only
patch a send site to an Open PIC if the send is a normal one.  So don't patch
in ceSend:above:to:numArgs:, and only if not a super send in ceSend:super:to:numArgs:.