CogVm source as per VMMaker.oscog-eem.2315
Double-back on the rash statement that there was a bad bug in ceSend:above:to:numArgs:
which was written to accept an association, not a class.  This was by design.
Generalize the JIT support for directed super sends so that it accepts both the
	pushLiteral: foo; directedSuperSend: bar numArgs: baz, and
	pushLiteralVariable: foo; directedSuperSend: bar numArgs: baz
forms.  Introduce aother directed send trampoline table for directedSuperBindingSends.
Implement ceSend:aboveClassBinding:to:numArgs: using ceSend:above:to:numArgs:.

Simplify, deleting genPushLiteralVariableGivenDirectedSuper: et al, and simply having
genPushLiteralIndex: peek for a directed super send if the bytecode set includes it.
Add the directedSendUsesBinding variable, set only by genPushLiteralIndex: and
consumed by genSendDirectedSuper:numArgs:, to distinguish between the two forms.

Keep the unused printPCMapPairsFor:, but mark it as doNotGenerate.
Implement targetMethodAndSendTableFor:annotation:into: in terms of
offsetAndSendTableFor:annotation:into:.