VMMaker.oscog-eem.2316.mcz Interpreter: Fix bad bug in reverseDisplayFrom:to: feedback from the leak checker.  The displayBits are not uopdated soon enough after a compaction and objects may be overwritten.  So refactor postGCAction: to extract postGCUpdateDisplayBits which is also used by reverseDisplayFrom:to: to obtain up-to-date bits. Interpreter Simulator: Fix snapshot write with a 64-bit memory. Make the globalSessionID (& thisSessionID in InterpreterProxy) a 32-bit quantity always.  Thse were using SmallInteger>>maxVal with the assumption that SmallIntegers are less than 32-bits.
Rewrite primitiveDeferDisplayUpdates to invoke indirectly ioForceDisplayUpdate, and implement ioForceDisplayUpdate to allow the display to refresh, hence curing the lack of display update in the StackInterpreter simulator, and allowing the removal of CogVMSimulator>>primitiveDeferDisplayUpdates.

Spur Simulator:
Speed up the leak checker by about a factor of two by removing the bounds checking versions of fetchPointer:ofObject: in the simulator subclasses for the duration, and by using bitShift: instead of << in addressAfter:

Socket Plugin SImulator:
Maintain the resolver's status correctly, instead of the broken deferal to the host's NetNameResolver.  This way the simulated image actually initializes itself.

SecurityPlugin simulator:
Answer true to ioCanRenameImage