CogVM source as per VMMaker.oscog-eem.2316
Interpreter:
Fix bad bug in reverseDisplayFrom:to: feedback from the leak checker.  The displayBits
are not uopdated soon enough after a compaction and objects may be overwritten.  So
refactor postGCAction: to extract postGCUpdateDisplayBits which is also used by
reverseDisplayFrom:to: to obtain up-to-date bits.

Rewrite primitiveDeferDisplayUpdates to invoke indirectly ioForceDisplayUpdate, and
implement ioForceDisplayUpdate to allow the display to refresh, hence curing the lack
of display update in the StackInterpreter simulator.