VMMaker.oscog-eem.2320.mcz Spur: Fix a bad bug in SpurPlnningCompactor.  unmarkObjectsFromFirstFreeObject, used when the compactor requires more than one pass due to insufficient savedFirstFieldsSpace, expects the corpse of a moved object to be unmarked, but copyAndUnmarkObject:to:bytes:firstField: only unmarked the target.  Unmarking the corpse before the copy unmarks both.  This fixes a crash with ReleaseBuilder class>>saveAsNewRelease when non-use of cacheDuring: creates lots of files, enough to push the system into the multi-pass regime.
Cogit:
Fix an assert fail during simulation with the openPICList.  Simply void the list when doing unlinkAllSends, and change removeFromOpenPICList: to neither crash nor complain when attempting to remove a PIC when the list is empty.  Have printOpenPICList answer the length of the list.

Simulation:
Add byte count text update to stack overflow in the cogit.