StackToRegisterMappingCogit: Get much stricter with simSpillBase for the benefit of RegisterAllocatingCogit (and eventually SistaCogit).
updateSimSpillBase is now called after the new descriptor is pushed and either counts up or counts down as required to guarantee it points one beyoind the last spilled entry.  The assertCorrectSimStackPtr asserts now insist on this.  With this change we can start to replace all the simSpillBase max: 0 prharses with simply simSpillBase.

Deal with the assett failures due to frameless methods with unused temporaries by simply refusing to JIT them; see the new methodFoundInvalidPostScan hook.