CogVM source as per VMMaker.oscog-eem.2397
Fix some compiler warnings in the interpreter (including unused variables).
Make methodReturnString: safe in the presence of allocation failures, but note
that, because it may fail on returning a result (which implies the primitive
has done its work) this is dubious at best.

Fix Slang not emitting variable declarations for 'extern ...' declarations.