RegisterAllocatingCogit. ensureReceiverResultRegContainsSelf should use a simple super send, since StackToRegisterMapping's version merely flushes the stack.  This results in poor code when assiging a method result to an inst var; the result comes back in ReceiverResultReg, which must be copied to another register to allow ReceiverResultReg to be loaded with self for the store.
Fix a simulation-only aliasing bug in restoreSimStackFromScratch.

Drop the obsolete noMustBeBoolean support in genJumpIf:to: and make sure to copy the sim stack to sratch /before/ any merge code is generated in the jumps.

Relax an assert in assertCorrectSimStackPtr.

Improve some comments for MNU and store trampolines.

Use the block inlining support to write some neater enumerators over the sim stack.

RegisterAllocatingCogit happily simulates a startreader-64 image and evaluates 3+4 & Smalltalk quit.

N.B. A weakness with the RegisterAllocatingCogit is that it does not currently map from registers to stack entries, hence if two ajoining sequences of bytecodes load the same register, but do so by pushing new desacriptors that are then popped afterwards, the register load will be repeated because there is no record that a particular register holds the value of a newly minted descriptor pushed on the stack.