VMMaker.oscog-AlistairGrant.2426 274: primitiveFileStdioHandles() fails to return nil if stdio file is not available
Modify FilePlugin>>primitiveFileStdioHandles to treat validMask = 0 as successfully determining that no stdio streams are available.

Previously validMask = 0 would result in the primitive failing with Unsupported.  However having no stdio streams available is normal on Windows, and so shouldn't signal an error.

sqFileStdioHandlesInto() can either return -1 to signal an unspecified error or use primitiveFailFor() or primitiveFailForOSError() to specify an error.