VMMaker.oscog-eem.2431.mcz Core VM: Kernel of support for failing FFI calls that raise exceptions.  This provides simulation support for catching the signal and activating the invoking method as a primitive failure.  The platform exception handlers still need of course to be modified to invoke this support.
Refactoring:
Move reenterInterpreter up to StackInterpreter from CoInterpreter, along with all relevant methods.  Add mustBeInterpreterFrame agument to justActivateNewMethod: to that the Cog VM can insist on activating the failing FFI invoking method in the interpreter, which simplifies the machinery in activateFailingPrimitiveMethod, which is the api entry-point that does the work of failing the primitive and long-jumping to the interpreter.  Fix a bug in CoInterpreter's justActivateNewMethod: which left the instructionPointer one ahead if activating a faiiing primitive method.  Add primitiveFailForFFIException:at: as a setter for the exception state.

Add PrimErrFFIException.  Update cloneOSErrorObj:numSlots: to deal with a three slot subclass of PrimitiveError that adds a pc at which an exception took place (exception can be represented by errorCode).

Plugins:
Have the ThreadedFFIPlugin create a valid stack allocation so it can simulate a callout.  Have its dispatchFunctionPointer:[with:*] methods catch MessageNotUnderstood and invoke primitiveFailForFFIException:at: and then activateFailingPrimitiveMethod.  Update morphIntoConcreteSubclass: to run any relevant class initialization if required.

Misc:
Initializing extensions should occur before fetching a bytecode, in case the bytecode is itself extended.

Add the 5 & 6 argument perform: methopds to this package, instead of relying on a file-in that ends up causing me to accidentally commit them to Kernel.

Have attemptToComputeTempNamesFor: use newFrom: instead of withAll: to work in Pharo.  Hopefully temporary.