Finish the VMMaker side of the failing of FFI callouts on exceptions.
Have the ThreadedFFIPlugin always call disownVM: and ownVM:, providing a flag that states if this is a threaded callout or not, and another flag that states that this is an FFI call.

Provide ffiExceptionResponse as a global variable the VM command line argument processors can set.  By default this is 0 and whether an FFI exception is caught or not depends on there being an error code in the FFI callout.  If there is one any exception will be caught; if there isn't a crash will occur as always.  If ffiExceptionResponse is < 0 FFI exceptions will never be caught and never delivered as primitive failures (which will be selected by -nofailonffiexception).  If ffiExceptionResponse is > 0 then FFI exceptions will always be caught and always delivered as primitive failures (which will be selected by -failonffiexception).

primitiveFailForFFIException:at: now fails if
- in an FFI call as indicated by DisownVMForFFICall being set in inFFIFlags
- ffiExceptionResponse > 0 or ffiExceptionResponse = 0 and newMethod has an error code

So it should always be called from the platform exception handlers such as sigsegv on Unix platforms.