CogVM source as per VMMaker.oscog-eem.2436
Core VM:
Support for failing FFI calls that raise exceptions.  primitiveFailForFFIException:at:
is the entry point for the platform exception handlers (sigsegv,
squeakExceptionHandler, et al) to catch the signal and call to activate the
invoking method as a primitive failure.

Have the ThreadedFFIPlugin always call disownVM: and ownVM:, providing a flag
that states if this is a threaded callout or not, and another flag that states
that this is an FFI call.

Provide ffiExceptionResponse as a global variable the VM command line argument
processors can set.  By default this is 0 and whether an FFI exception is caught
or not depends on there being an error code in the FFI callout.  If there is one
any exception will be caught; if there isn't a crash will occur as always.  If
ffiExceptionResponse is < 0 FFI exceptions will never be caught and never
delivered as primitive failures (which is selected by -nofailonffiexception).
If ffiExceptionResponse is > 0 then FFI exceptions will always be caught and
always delivered as primitive failures (which is selected by -failonffiexception).

primitiveFailForFFIException:at: now fails if
- in an FFI call as indicated by DisownVMForFFICall being set in inFFIFlags
- ffiExceptionResponse > 0 or ffiExceptionResponse = 0 and newMethod has an error code

So it should always be called from the platform exception handlers such as sigsegv on Unix platforms.

Refactoring:
Move reenterInterpreter up to StackInterpreter from CoInterpreter, along with
all relevant methods.  Add mustBeInterpreterFrame agument to justActivateNewMethod:
to that the Cog VM can insist on activating the failing FFI invoking method in
the interpreter, which simplifies the machinery in activateFailingPrimitiveMethod,
which does the work of failing the primitive and long-jumping to the interpreter.
Fix a bug in CoInterpreter's justActivateNewMethod: which left the
instructionPointer one ahead if activating a failing primitive method.
Add primitiveFailForFFIException:at: as a setter for the exception state.

Add PrimErrFFIException.  Update cloneOSErrorObj:numSlots: to deal with a three
slot subclass of PrimitiveError that adds a pc at which an exception took place
(exception can be represented by errorCode).

Misc:
Initializing extensions should occur before fetching a bytecode, in case the bytecode is itself extended.

FileAttributesPlugin:

274: Update primitiveFileStdioHandles error handling.

Previously validMask = -1 was considered an error.  Now validMask < 0 is
considered an error and the value is returned to the image with
primitiveFailForOSError().

This is more robust against word length in the VM and provides a general
mechanism for the plugin to provide error information back to the image.

274: primitiveFileStdioHandles() fails to return nil if stdio file is not
available

Modify FilePlugin>>primitiveFileStdioHandles to treat validMask = 0 as
successfully determining that no stdio streams are available.

Previously validMask = 0 would result in the primitive failing with Unsupported.
However having no stdio streams available is normal on Windows, and so shouldn't
signal an error.

sqFileStdioHandlesInto() can either return -1 to signal an unspecified error or use primitiveFailFor() or primitiveFailForOSError() to specify an error.

Image Directory:
Add LoadFFI.st to ease creating spurreader images containing the FFI.
Fix the use of getGoodSpur[64]VM.sh so the buildspurtrunkreader[64]image.sh
commands can take both a -vm foo and an FFI argument.
Nuke Object-performwithwithwithwithwith.st.  This is now in VMMaker.oscog.