CogVM source as per VMMaker.oscog-eem.2445
BitBlt plugin copyBits.
Fix degenerate calculations of preload and skew (i.e. a preload that sets
notSkewMask to all ones and skewMask to zero, and there-by fix accessing the
word beyond the end of a bitmap.  If using external forms such access can crash
the VM by trying to access a word that is not in memory (e.g. in an unmapped
page).  N.B. when preload is true, notSkewMask is all ones and skewMask is zero
this extra word is read but discarded.

Clean up primitiveCopyBits & primitiveWarpBits to use the more modern
(and simpler) methodReturnReceiver style.