CogVM source as per VMMaker.oscog-eem.2453
General:
Fix error code reporting for primitiveAdoptInstance.  It uses changeClassOf:to:
which expects the class to be the argument, so BadArgument/BadReceiver errors
must be swapped.

Spur:
Simplify the two changeClassOf:to: implementations, fixing a bug when a byte
class adopts a 32 bit indexable instance on the 32 bit system.

Spur Cogit:
Fix a regression in 32 bit genGetSizeOf:into:formatReg:scratchReg:abortJumpsInto:
that failed to abort for 64 bit indexable receivers.

Eliminate unnecessary branch in 32 bit version of genPrimitiveAtSigned:, to
match the 64 bit version.

Plugins:
Eliminate cCode: usage in the B3DAcceleratorPlugin and HostWindowPlugin using
the new "var args" style.  Rewrite mem:mo:ve: et al in the new style.