VMMaker.oscog-AlistairGrant.2455 VMClass strlen, strncpy and getenv
Pharo stores UTF8 encoded strings in ByteArrays (ByteString, strictly speaking, expects to only store characters that can be represented as a single byte in UTF8).  ByteArrays are also used within the simulator to represent buffers allocated by the simulator.  As such, the strings may either be the length of the ByteArray or less than the ByteArray size and null terminated.

These changes extend strlen: and strncpy:_:_: to handle ByteArrays and add some tests (tests for strings in the object memory are todo).

InterpreterPrimitives>>primitiveGetenv: returned nil rather than 0 in the simulator when a variable that isn't defined is requested.