Plugins: BitBltSimulation Fix a bug where the copyBits primitive read past the end of the source bitmap.  sourceSkewAndPointerInit would sometimes compute preload true and skew truye before then setting skew to -32.  There was a hack fix in copyLoop, but the real bug in copyLoop was the inner loop special case for rule 3 (over, or STORE).  It must fall back to the loop beginning "thisWord := self srcLongAt: sourceIndex." when preload is false, only using the loop beginning "self dstLongAt: destIndex put: prevWord." when preload and: [hDr = 1] is true.  Simplify the preload calculation in sourceSkewAndPointerInit by computing a mask similar to mask1 and comparing it against mask1 to see that no bits would be lost (if the mask computed for source is larger than mask1 then a preload is necessary).  Hence simplify the unskew and skewMask setup in copyLoop.
Improve the performance of primitiveDisplayString by pulling the lockSurfaces and unlockSurfaces implicit in copyBits out of the loop and replacing copytBits with copyBitsLockedAndClipped in the loop.

Fix primitivePixelValueAt so that it simulates correctly.

Other plugins:
Remove null implementations of initialiseModule and/or shutdownModule.  The loader/unloader nvokes these only if they exiust, consequently null versions are simply a waste of time and space.

Simulation:
Add support for computing the hash of the destination bitmap after copyBits in BitBltSimulator which was key in identifying the cases for which the above BitBltSimulation>>copyLoop bug fix applied.

Rewrite JPEGReadWriter2Plugin so that it simulates.  This is now a nice example of a recursive SmartSyntaxPlugin; the simulation of the plugin relies on the real implementation of the plugin in the underlying VM.

Fix InterpreterProxy>>byteSizeOf: for 64-bits and for bits objects.  Implement CArray>>asByteArray & CArray>>overwriteContentsWith: to ease copying bytes to/from the heap when simulating primitives (key for the JPEGReadWriter2Plugin simulation support).

Fix the return value of InterpreterProxy halt during simulation.