CogVM source as per VMMaker.oscog-eem.2463
Plugins:
BitBltSimulation
Fix a bug where the copyBits primitive read past the end of the source bitmap.
sourceSkewAndPointerInit would sometimes compute preload true and skew true
before then setting skew to -32.  There was a hack fix in copyLoop, but the real
bug in copyLoop was the inner loop special case for rule 3 (over, or STORE). It
must fall back to the loop beginning "thisWord := self srcLongAt: sourceIndex."
when preload is false, only using the loop beginning "self dstLongAt: destIndex
put: prevWord." when preload and: [hDr = 1] is true.

Simplify the preload calculation in sourceSkewAndPointerInit by computing a mask
similar to mask1 and comparing it against mask1 to see that no bits would be
lost (if the mask computed for source is larger than mask1 then a preload is
necessary).  Hence simplify the unskew and skewMask setup in copyLoop.

Improve the performance of primitiveDisplayString by pulling the lockSurfaces
and unlockSurfaces implicit in copyBits out of the loop and replacing copyBits
with copyBitsLockedAndClipped in the loop.

Fix primitivePixelValueAt so that it simulates correctly.

Other plugins:
Remove null implementations of initialiseModule and/or shutdownModule.  The
loader/unloader invokes these only if they exist, consequently null versions
are simply a waste of time and space.