VMMaker.oscog-eem.2477 Plugins: Fix primDigitCompare for SmallIntegers.  The old code compared the values of the receiver and argument, not their magnitudes, if both were SmallIntegers.
Simplify the FilePlugin. removing an unnecessary indirection around dir_Delimitor, and eliminating an unnecessary cCode:inSmalltalk:.

Simulator:
Fix a regression with FilePlugin loading due to the simplifications to simulated plugin loading introduced in VMMaker.oscog-eem.2476.  mappedPluginEntries must be extended, not assigned to.

Disable simulation for ClipboardExtendedPlugin.  This would be a good exercise for someone who wanted to try writing a plugin simulation themselves, hint, hint.