CogVM source as per VMMaker.oscog-eem.2488
StackInterpreter:
Fix a bug where a reference in a married context in a base frame would prevent
garbage collection.  The same issue is fixed for normal marriage/divorce of
contexts, but was not handled in makeBaseFrameFor:.  Thanks to Ryan Macnak for
identifying both bug and fix.

SelectiveCompactor just abort compaction when it fails to find a segment to
compact into or to allocate one, instead of crashing the VM with the error
'no segment to compact into'

Sista:
Make non-local return safe in the presence of Sista outer-context-less blocks;
i.e. throw cannotReturn: when a home context can't be found instead of crashing.