MacOS X does support the tcsetattr serial port support supposed by sqUnixSerial.c.  So instead of using the null support that was in sqMacSerialPort.c, update the Mac builds to use sqUnixSerial.c, and rename the mac's file to platforms/Cross/plugins/SerialPlugin/sqNullSerialPort.c
THIS IS UNTESTED!  In  particular the code in sqUnixSerial.c (dubiously IMO)
does not use serialPortBaseName, and instead makes an assumption that
serial ports are as defined by serialPortBaseNameDefault, i.e. are in the range
/dev/tty50 to /dev/tty59.  It would be great if
a) someone tested the code on MacOS X
b) the naming convention were better.
On Mac OS X (see e.g. https://software.intel.com/en-us/setting-up-serial-terminal-on-system-with-mac-os-x),
the serial ports are /dev/tty.NAME., e.g. /dev/cu.Bluetooth-Incoming-Port,
/dev/cu.BoseQuietControl30-SPPD, etc.