Make Xcode project to build iOS 64bit ARM app using a Cuis5.0-3081-spur-64 image.
noteworthy hacks:

A hard coded change to gcc3x-interp.c to add jumpTable & checkAllocFiller to the static structure, this seems to improve the stack vm performance, cross checking is required.
Some issues with a C inferred defines versus explicit.

This code base excluding a image/change is used to support an app in the app store.