Check sanity of curled gzip data
The getGoodSpur*VM.sh scripts assume that the VM mentioned in the
topmost notification, is always there on bintray.  This is not
always the case.  Unfortunately, bintray will not return a 404 code
on a 404 error, but instead give a 200 with garbage data.  While we
can't do anything about that, at least make an attempt to bail if
the file doesn't even look like gzip-compressed data.