Fix sqWin32Security.c UNICODE problems
We cannot pass a TCHAR * to a function expecting a WCHAR *, or the compiler logically barks.
`isAccessiblePathName` takes a WCHAR * parameter
`fromSqueak` answers a TCHAR *, which is currently a char * (because UNICODE is undefined).

We want to be able to query internationalized path name, then the best (portable) thing to do is stick to UTF-8.
We thus now interpret the pathName/fileName as UTF8-encoded in the authorization query primitives.

NOTE: no effort is attempted to handle long path names here.