Use GetLocaleInfoA rather than generic version
This is the least change that makes the plugin compile correctly with -DUNICODE

But this is maybe not the right thing.
If the answer contains a character that is not available in the code page,
then it will be replaced by a question mark.
https://docs.microsoft.com/en-us/windows/desktop/api/winnls/nf-winnls-getlocaleinfow

The right thing would be to go thru GetLocaleInfoW and WideCharToMultiByte dance to answer an UTF-8 encoded string.