CogVM source as per VMMaker.oscog-eem.2504/FileAttributesPlugin.oscog-akg.49
Slang:
Fix a bug in type inferrence that resulted in the inferred return type of
findMapLocationForMcpc:inMethod: to flip between usqInt (correct) and sqInt
(incorrect).  addTypesFor:to:in: needed to answer if it was inferring a return
type from an untyped variable, as well as an untyped method.

Plugins: FileAttributesPlugin v2.0.8

Fix some of the doits in Slang Test Workspace.text