pharo: Forward the packaging directory from pharo-vm here
Add structure and rules to build a source package for Debian/Ubuntu
and use it for RPM based distributions. A source package can be built
by the Open Build Service (OBS) or by users. It is intended to be
rebuildable by any user.

Next steps are to hook this into the travis-ci build to upload a
new source package on every build into an experimental feed.