CogVM source as per VMMaker.oscog-eem.2523
Spur:
Fix image segment storage when there is insufficient contguous space to store
the array of objects to be included in a segment.  I wrote the code to answer a
suitable error object (a SmallInteger of the required size), but forgot to
handle the return, hence causing horrible crashes as the VM attempted to access
objects in a SmallInteger.  this was Max Leske's failing segment storage test
case (a 66.5Mb segment).

Handle the case by introducing a new error code (PrimErrNeedCompact) and having
the primitive perform a full GC when it gets the error code, and then to retry
segment storage.

Have printFreeTree/printFreeChunk:printAsTreeNode: mark the root node with a '+'

Fix some slips in free chunk integrity checking.