CogVM source as per VMMaker.oscog-eem.2530
Slang: Fix bug in dead-code elimination.  In checking whether a name is defined
at compile-time or not we have to deal with both TDefineNode and TVarableNode,
and so must use nameOrValue, not just value.  This affects finalizeReference:
which was having variables used in a "cppIf: PharoVM ifTrue: ..." arm deleted
since the code generator incorresctly assessed the arm as dead code.