debug: Prepare to let the app own stderr
The application should own stdout/stderr. This is especially important
when building CLI applications (like pillar, ecstatic, etc) and having
none Smalltalker use them.

Introduce a VM_ERR() function to return a shared FILE* that can be
used for error logging.

For plugins we will need to extend VirtualMachine to carry a pointer
to the VM_ERR method.