CogVM source as per VMMaker.oscog-eem.2536
Have primitiveNextObject fail in Spur.  DTL's recent Kernel changes are good,
but it's still preferrable for the primtiive to check since performance isn't
an issue in Spur (we use allObjects).

Have the code generator eliminate redundant assignments (aVar = aVar) which tend
to be generated from Smalltalk idioms where a variable is assigned from a method
taking it as a parameter that in some circumstances is a noop, e.g. assignment
to latestContinuation from maybeDealWithUnsafeJumpForDescriptor:pc:latestContinuation:.
This to reduce C compiler warnings.